Serialize all wasi-types into one lib.rs file, adjust Oflags
