+++ lib/wasi-types-generated/regenerate.sh
@@ -14,5 +14,6 @@ wit-bindgen rust-wasm \
     --out-dir "$BASEDIR"/src/wasi
 
 awk '{sub(/mod output/,"pub mod output")}1' src/wasi/bindings.rs > src/wasi/bindings2.rs
+cargo fmt --all
 cp src/wasi/bindings2.rs src/wasi/bindings.rs
 rm src/wasi/bindings2.rs
